Working on patch 1.1
Fixing a bug where only parts of folders are included in file lists, hence not all files being copied over. Fixed the issue for copying entire drives from the root, but the issue persists for specific folders. The query for folders currently only lists files and folders directly under the parent folder, while it needs to list all the files and folders with the parent folder id in the 'parents' collection attribute. A complete fix will be coming in the next few days.