Folder replication removed and other issues fixed
Full Drive replication/backup is fixed and can be used, folder replication has been removed temporarily. It will be fixed somewhere next week. The files that are not children of the target folder just need to be filtered out of the data frame. Other than that, the program is fully functional right now.